split match nodes to separate file
