+++ tests/TestResults/match-by-context.test.ts
@@ -1,3 +1,4 @@
+jest.unmock('../../src/TestResults/MatchNode');
 jest.unmock('../../src/TestResults/match-by-context');
 jest.unmock('../test-helper');
